requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
environs==9.3.1
fsspec==0.8.7
future==0.18.2
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
imgkit==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.3.4
monai==0.4.0
multidict==5.1.0
nibabel==3.2.1
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytorch-lightning==1.2.5
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.13.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
timm==0.4.5
torch==1.8.0
torchmetrics==0.2.0
torchvision==0.9.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3